requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==19.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
discord.py==2.1.0
ez-setup==0.9
filelock==3.8.0
frozenlist==1.3.3
giphy-client==1.0.0
huggingface-hub==0.11.0
idna==2.9
multidict==4.7.6
numpy==1.23.4
packaging==21.3
peewee==3.13.3
pycparser==2.20
pydantic==1.5.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.13.0
PyYAML==6.0
regex==2022.10.31
requests==2.24.0
six==1.15.0
#tokenizers==0.13.2
#torch==1.13.0
tqdm==4.64.1
#transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
httpx==0.23.0
beautifulsoup4==4.11.1
youtube-dl==2021.12.17
ytdl==1.0.0rc5
git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074